refactor: extract product movement report components for better maintainability
- Split ProductMovement page into separate components:
  - FastMovingTable: Displays fast-moving products
  - SlowMovingTable: Displays slow-moving products
  - SummaryCards: Shows summary statistics
- Reduce main page component complexity
