+++ tests/Feature/Report/ProductMovementControllerTest.php
@@ -39,17 +39,17 @@ function movementSale(User $user, Product $product, int $qty = 1, array $saleAtt
 
 function movementAdmin(): User
 {
-    return User::where('role', 'admin')->first() ?? User::factory()->create(['role' => 'admin']);
+    return User::role('admin')->first() ?? User::factory()->create(['role' => 'admin']);
 }
 
 function movementCashier(): User
 {
-    return User::where('role', 'cashier')->first() ?? User::factory()->create(['role' => 'cashier']);
+    return User::role('cashier')->first() ?? User::factory()->create(['role' => 'cashier']);
 }
 
 function movementWarehouse(): User
 {
-    return User::where('role', 'warehouse')->first() ?? User::factory()->create(['role' => 'warehouse']);
+    return User::role('warehouse')->first() ?? User::factory()->create(['role' => 'warehouse']);
 }
 
 // ============================================================
